v0.1.0
Initial release. Spinnaker can do simple page tracking, and some serving
of simple information. In addition to the number of visits each page
gets, it also tracks when each visit occured, as well as some meta info
about the page (if it's the index or a feed). There are only 4 API
queries: /, /today, /week, /month; which each return page info for the
their respective timeframes (week = this week, month = this month, / =
the last 24 hours). Spinnaker uses two tables to store the information,
and will do so using DATABASE_URL (or a sqlite3 database if that
environment variable is missing). The API endpoints are protected via
the 'listen' configuration, which defaults to '127.0.0.1'.